Update service-fabric-cluster-creation-via-portal.md
The link for this content is broken:
"c. The minumum number of VMs for the primary node type is driven by the reliablity tier you choose. The default for the reliablity tier is Silver. Read more on how to [choose the Service Fabric cluster reliability and durability](service-fabric-cluster-reliability-and-durability.md) document."